+++ cli/command_repository_upgrade.go
@@ -156,6 +156,7 @@ func (c *commandRepositoryUpgrade) validateAction(ctx context.Context, rep repo.
 		}
 	}
 	if err == nil {
+		log(ctx).Infof("success: upgraded index matches old index")
 		return nil
 	}
 
@@ -408,11 +409,6 @@ func (c *commandRepositoryUpgrade) upgrade(ctx context.Context, rep repo.DirectR
 	// update format-blob and clear the cache
 	if err := rep.FormatManager().SetParameters(ctx, mp, blobCfg, rf); err != nil {
 		return errors.Wrap(err, "error setting parameters")
-	}
-
-	// poison V0 index so that old readers won't be able to open it.
-	if err := content.WriteLegacyIndexPoisonBlob(ctx, rep.BlobStorage()); err != nil {
-		log(ctx).Errorf("unable to write legacy index poison blob: %v", err)
 	}
 
 	// we need to reopen the repository after this point
