+++ repo/format/upgrade_lock.go
@@ -12,7 +12,11 @@ import (
 // BackupBlobIDPrefix is the prefix for all identifiers of the BLOBs that
 // keep a backup copy of the FormatBlobID BLOB for the purposes of rollback
 // during upgrade.
-const BackupBlobIDPrefix = "kopia.repository.backup."
+const (
+	BackupBlobIDPrefix = "kopia.repository.backup."
+
+	LegacyIndexPoisonBlobID = "n00000000000000000000000000000000-repository_unreadable_by_this_kopia_version_upgrade_required"
+)
 
 // BackupBlobID gets the upgrade backu pblob-id fro mthe lock.
 func BackupBlobID(l UpgradeLockIntent) blob.ID {
@@ -71,6 +75,18 @@ func (m *Manager) SetUpgradeLockIntent(ctx context.Context, l UpgradeLockIntent)
 	return m.repoConfig.UpgradeLock.Clone(), nil
 }
 
+// WriteLegacyIndexPoisonBlob writes a "poison blob" that will prevent old kopia clients
+// that have not been upgraded from being able to open the repository after its format
+// has been upgraded.
+func WriteLegacyIndexPoisonBlob(ctx context.Context, st blob.Storage) error {
+	//nolint:wrapcheck
+	return st.PutBlob(
+		ctx,
+		LegacyIndexPoisonBlobID,
+		gather.FromSlice([]byte("The format of this repository has been upgraded and cannot be read by old clients")),
+		blob.PutOptions{})
+}
+
 // CommitUpgrade removes the upgrade lock from the from the repository format
 // blob. This in-effect commits the new repository format t othe repository and
 // resumes all access to the repository.
@@ -84,6 +100,11 @@ func (m *Manager) CommitUpgrade(ctx context.Context) error {
 
 	if m.repoConfig.UpgradeLock == nil {
 		return errors.New("no upgrade in progress")
+	}
+
+	// poison V0 index so that old readers won't be able to open it.
+	if err := WriteLegacyIndexPoisonBlob(ctx, m.blobs); err != nil {
+		log(ctx).Errorf("unable to write legacy index poison blob: %v", err)
 	}
 
 	// restore the old format version
